quiche-client: add --send-priority-update param
quiche-client has supported two methods of sending extensible
priorities:

  1) via a header, controlled by the `-H` parameter. This applies to all
     requests, if more than one URL (or the `-n` parameter) was
     provided.
  2) via a custom querystring format such as `?u=1&i=1`. This format is
     understood by quiche-server, facilitating experimentation with
     priorities by supplying multiple URLs with individual priorities.

With this change, we enhance the second method with the optional
`--send-priority-update` parameter. When it is supplied, quiche-client
will use the querystring as input to construct an individual
PRIORITY_UPDATE frame for the request.

quiche-server has been updated to reflect this change. It now attempts
to take PRIORITY_UPDATE frames when building a response, and sets the
response priority accordingly.